requirements.txt
# Core dependencies (pinned for reproducibility)
click==8.3.1
PyYAML==6.0.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
pydantic==2.12.5
pydantic_core==2.41.5

# Tool server
uvicorn==0.41.0
starlette==0.52.1
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
anyio==4.12.1

# Repo packaging
zstandard==0.25.0

# Dev / test
pytest==9.0.2
pytest-timeout==2.4.0

# Transitive (pinned)
annotated-types==0.7.0
attrs==25.4.0
iniconfig==2.3.0
pluggy==1.6.0
pygments==2.19.2
referencing==0.37.0
rpds-py==0.30.0
typing-extensions==4.15.0
typing-inspection==0.4.2
